pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ats-score-resume"
version = "0.1.0"
description = "Aplicacao para avaliar curriculos com metricas ATS e aderencia a vagas."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.12,<5",
    "pypdf>=5,<6",
    "python-docx>=1.1,<2",
    "requests>=2.32,<3",
    "streamlit>=1.43,<2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3,<9",
]

[tool.pytest.ini_options]
addopts = "-p no:cacheprovider"
pythonpath = ["src"]
testpaths = ["tests"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
